requirements.txt
llama-index==0.11.13
llama-index-indices-managed-genesis==0.2.2
llama-index-agent-llm-compiler==0.2.0
llama-index-agent-openai==0.3.4
llama-index-llms-openai==0.2.9
llama-index-llms-anthropic==0.3.1
llama-index-llms-together==0.2.0
llama-index-llms-groq==0.2.0
llama-index-llms-fireworks==0.2.0
llama-index-llms-cohere==0.3.0
llama-index-llms-gemini==0.3.5
llama-index-tools-yahoo-finance==0.2.0
llama-index-tools-arxiv==0.2.0
llama-index-tools-database==0.2.0
llama-index-tools-google==0.2.0
llama-index-tools-tavily_research==0.2.0
openinference-instrumentation-llama-index==3.0.2
opentelemetry-proto==1.16.0
arize-phoenix-otel==0.5.1
tokenizers>=0.20
pydantic==2.9.2
retrying==1.3.4
pymongo==4.6.3
python-dotenv==1.0.1
tiktoken==0.7.0
dill==0.3.8
